examples/embedding/telegram-bot/requirements.txt
python-telegram-bot[ext]>=21.0
httpx>=0.27
